+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 next
 ----
 
-- Support for CTEs (`WITH ...` clauses) in `DELETE FROM` queries
+- Support for CTEs (`WITH ...` clauses) in `DELETE FROM` and `UPDATE` queries
 - `DeleteFrom`, `InsertInto` and `Update` structs now do not implement `Default`, because such empty struts are not valid
 - All query structs now implement `Clone` so you can, well, clone them
 - All returned query structs are marked as `must_use`